Move process_blockvss to indepenedent file
