+++ src/signer_node/message_processor/mod.rs
@@ -5,7 +5,9 @@ use crate::rpc::TapyrusApi;
 mod process_candidateblock;
 mod process_completedblock;
 mod process_nodevss;
+mod process_blockvss;
 
 pub use process_candidateblock::process_candidateblock;
 pub use process_completedblock::process_completedblock;
 pub use process_nodevss::process_nodevss;
+pub use process_blockvss::process_blockvss;
