+++ src/signer_node/mod.rs
@@ -33,6 +33,7 @@ use crate::util::*;
 use crate::signer_node::message_processor::process_candidateblock;
 use crate::signer_node::message_processor::process_completedblock;
 use crate::signer_node::message_processor::process_nodevss;
+use crate::signer_node::message_processor::process_blockvss;
 
 /// Round interval.
 pub static ROUND_INTERVAL_DEFAULT_SECS: u64 = 60;
@@ -317,20 +318,7 @@ impl<T: TapyrusApi, C: ConnectionManager> SignerNode<T, C> {
             MessageType::Candidateblock(block) => process_candidateblock(&message.sender_id, &block, self),
             MessageType::Completedblock(block) => process_completedblock(&message.sender_id, &block, self),
             MessageType::Nodevss(vss, secret_share) => process_nodevss(&message.sender_id, vss, secret_share, self),
-            MessageType::Blockvss(
-                blockhash,
-                vss_for_positive,
-                secret_share_for_positive,
-                vss_for_negative,
-                secret_share_for_negative,
-            ) => self.process_blockvss(
-                blockhash,
-                vss_for_positive,
-                secret_share_for_positive,
-                vss_for_negative,
-                secret_share_for_negative,
-                message.sender_id,
-            ),
+            MessageType::Blockvss(blockhash, vss_for_positive, secret_share_for_positive, vss_for_negative, secret_share_for_negative) => process_blockvss(&message.sender_id, blockhash, vss_for_positive, secret_share_for_positive, vss_for_negative, secret_share_for_negative, self),
             MessageType::Blocksig(blockhash, gamma_i, e) => {
                 self.process_blocksig(blockhash, gamma_i, e, message.sender_id)
             }
@@ -377,186 +365,6 @@ impl<T: TapyrusApi, C: ConnectionManager> SignerNode<T, C> {
             }
         };
         next_state
-    }
-
-    fn process_blockvss_inner(
-        &self,
-        blockhash: Hash,
-        shared_block_secrets: &BidirectionalSharedSecretMap,
-    ) -> Option<(bool, SharedKeys)> {
-        let params = self.sharing_params();
-        log::trace!(
-            "number of shared_block_secrets: {:?}",
-            shared_block_secrets.len()
-        );
-        let block_opt: Option<Block> = match &self.current_state {
-            NodeState::Master {
-                candidate_block, ..
-            } => Some(candidate_block.clone()),
-            NodeState::Member {
-                candidate_block, ..
-            } => candidate_block.clone(),
-            _ => None,
-        };
-        if let Some(block) = block_opt.clone() {
-            if block.sighash() != blockhash {
-                log::error!("Invalid blockvss message received. Received message is based different block. expected: {:?}, actual: {:?}", block.sighash(), blockhash);
-                return None;
-            }
-        } else {
-            // Signer node need to receive candidateblock before receiving VSS.
-            log::error!("Invalid blockvss message received. candidateblock was not received in this round yet, but got VSS.");
-            return None;
-        }
-        if shared_block_secrets.len() == self.params.pubkey_list.len() {
-            let shared_keys_for_positive = Sign::verify_vss_and_construct_key(
-                &params,
-                &shared_block_secrets.for_positive(),
-                &(self.params.self_node_index + 1),
-            )
-            .expect("invalid vss");
-
-            let result_for_positive = Sign::sign(
-                &shared_keys_for_positive,
-                &self.priv_shared_keys.clone().unwrap(),
-                block_opt.clone().unwrap().sighash(),
-            );
-
-            let shared_keys_for_negative = Sign::verify_vss_and_construct_key(
-                &params,
-                &shared_block_secrets.for_negative(),
-                &(self.params.self_node_index + 1),
-            )
-            .expect("invalid vss");
-            let result_for_negative = Sign::sign(
-                &shared_keys_for_negative,
-                &self.priv_shared_keys.clone().unwrap(),
-                block_opt.clone().unwrap().sighash(),
-            );
-
-            let p = BigInt::from_str_radix(
-                "FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFEFFFFFC2F",
-                16,
-            )
-            .unwrap();
-            let is_positive = jacobi(&shared_keys_for_positive.y.y_coor().unwrap(), &p) == 1;
-            let (shared_keys, result) = if is_positive {
-                (shared_keys_for_positive, result_for_positive)
-            } else {
-                (shared_keys_for_negative, result_for_negative)
-            };
-
-            match result {
-                Ok(local_sig) => {
-                    self.connection_manager.broadcast_message(Message {
-                        message_type: MessageType::Blocksig(
-                            block_opt.clone().unwrap().sighash(),
-                            local_sig.gamma_i,
-                            local_sig.e,
-                        ),
-                        sender_id: self.params.signer_id,
-                        receiver_id: None,
-                    });
-                }
-                _ => (),
-            }
-            return Some((is_positive, shared_keys));
-        } else {
-            return None;
-        }
-    }
-    fn process_blockvss(
-        &mut self,
-        blockhash: Hash,
-        vss_for_positive: VerifiableSS,
-        secret_share_for_positive: FE,
-        vss_for_negative: VerifiableSS,
-        secret_share_for_negative: FE,
-        from: SignerID,
-    ) -> NodeState {
-        match &self.current_state {
-            NodeState::Master {
-                block_key,
-                shared_block_secrets,
-                candidate_block,
-                signatures,
-                round_is_done: false,
-                ..
-            } => {
-                let mut new_shared_block_secrets = shared_block_secrets.clone();
-                new_shared_block_secrets.insert(
-                    from,
-                    (
-                        SharedSecret {
-                            vss: vss_for_positive.clone(),
-                            secret_share: secret_share_for_positive,
-                        },
-                        SharedSecret {
-                            vss: vss_for_negative.clone(),
-                            secret_share: secret_share_for_negative,
-                        },
-                    ),
-                );
-                let shared_keys = self.process_blockvss_inner(blockhash, &new_shared_block_secrets);
-
-                match shared_keys {
-                    Some(keys) => NodeState::Master {
-                        block_key: block_key.clone(),
-                        shared_block_secrets: new_shared_block_secrets,
-                        block_shared_keys: Some((keys.0, keys.1.x_i, keys.1.y)),
-                        candidate_block: candidate_block.clone(),
-                        signatures: signatures.clone(),
-                        round_is_done: false,
-                    },
-                    None => NodeState::Master {
-                        block_key: block_key.clone(),
-                        shared_block_secrets: new_shared_block_secrets,
-                        block_shared_keys: None,
-                        candidate_block: candidate_block.clone(),
-                        signatures: signatures.clone(),
-                        round_is_done: false,
-                    },
-                }
-            }
-            NodeState::Member {
-                block_key,
-                shared_block_secrets,
-                candidate_block,
-                ..
-            } => {
-                let mut new_shared_block_secrets = shared_block_secrets.clone();
-                new_shared_block_secrets.insert(
-                    from,
-                    (
-                        SharedSecret {
-                            vss: vss_for_positive.clone(),
-                            secret_share: secret_share_for_positive,
-                        },
-                        SharedSecret {
-                            vss: vss_for_negative.clone(),
-                            secret_share: secret_share_for_negative,
-                        },
-                    ),
-                );
-                let shared_keys = self.process_blockvss_inner(blockhash, &new_shared_block_secrets);
-
-                match shared_keys {
-                    Some(keys) => NodeState::Member {
-                        block_key: block_key.clone(),
-                        shared_block_secrets: new_shared_block_secrets,
-                        block_shared_keys: Some((keys.0, keys.1.x_i, keys.1.y)),
-                        candidate_block: candidate_block.clone(),
-                    },
-                    None => NodeState::Member {
-                        block_key: block_key.clone(),
-                        shared_block_secrets: new_shared_block_secrets,
-                        block_shared_keys: None,
-                        candidate_block: candidate_block.clone(),
-                    },
-                }
-            }
-            _ => self.current_state.clone(),
-        }
     }
 
     fn process_blocksig(
